linux compiler must be happy

+++ SlopEngine/engine/handmade.cpp
@@ -769,7 +769,7 @@ internal void ScanFill(game_offscreen_buffer *Buffer, coordinate a, coordinate b
 
 	//Splits A triangle into two parts, top part (from y_top to y_mid), fills pixels within it
 	for(uint32 scan_index = 0; scan_index < uint32(y_top - y_mid); ++scan_index){
-		real32 current_y = bound_y_max.y - scan_index;
+		real32 current_y = bound_y_max.y - (real32)scan_index;
 
 		real32 x_pos_1 = x_at_y(bound_y_mid, bound_y_max, slope_short_top, current_y);
 		real32 x_pos_2 = x_at_y(bound_y_min, bound_y_max, slope_long, current_y);
@@ -784,14 +784,14 @@ internal void ScanFill(game_offscreen_buffer *Buffer, coordinate a, coordinate b
 		}
 		uint32 pixel_counter = 0;
 		while ( pixel_counter <= (uint32)RoundReal32ToInt32(x_pos_right - x_pos_left)){
-			DrawPixel(Buffer, RoundReal32ToInt32(x_pos_left)+ pixel_counter, RoundReal32ToInt32(current_y), color);
+			DrawPixel(Buffer, RoundReal32ToInt32(x_pos_left)+ (int32)pixel_counter, RoundReal32ToInt32(current_y), color);
 			pixel_counter++;
 		}
 	}	
 
 	//2nd triangle part from y_mid to y_min
 	for(uint32 scan_index = 0; scan_index < uint32(y_mid - y_bot); ++scan_index){
-		real32 current_y = bound_y_mid.y - scan_index;
+		real32 current_y = bound_y_mid.y - (real32)scan_index;
 
 		real32 x_pos_1 = x_at_y(bound_y_mid, bound_y_min, slope_short_bot, current_y);
 		real32 x_pos_2 = x_at_y(bound_y_min, bound_y_max, slope_long, current_y);
@@ -806,7 +806,7 @@ internal void ScanFill(game_offscreen_buffer *Buffer, coordinate a, coordinate b
 		}
 		uint32 pixel_counter = 0;
 		while ( pixel_counter <= (uint32)RoundReal32ToInt32(x_pos_right - x_pos_left)){
-			DrawPixel(Buffer, RoundReal32ToInt32(x_pos_left)+ pixel_counter, RoundReal32ToInt32(current_y), color);
+			DrawPixel(Buffer, RoundReal32ToInt32(x_pos_left)+ (int32)pixel_counter, RoundReal32ToInt32(current_y), color);
 			pixel_counter++;
 		}
 	}
